Clarifying the 'Developing with The Core SDK" section.
Motivation:
Currently it is not obvious for a new comer to this project to know
exactly which files to symlink and how to start the server locally. This
commit hopes to clarify this.

Modifications:
Added examples of creating the symbolic links.
Addes an example of starting the app locally and showing how the --url
option can be used.

Result:
Should be easier to follow the instructions.